Project-2/backend/requirements.txt
flask==2.3.0
langchain==0.1.0
google-generativeai==0.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.0
sqlalchemy==2.0.0
